rename announcements -> news_items
