--- database/tables/announcement_notifications.pg
@@ -1,12 +1,0 @@
-columns
-    announcement_id: bigint not null
-    id: bigint not null default nextval('announcement_notifications_id_seq'::regclass)
-    user_id: bigint not null
-
-indexes
-    announcement_notifications_pkey: PRIMARY KEY (id) USING btree (id)
-    announcement_notifications_user_id_announcement_id_key: UNIQUE (user_id, announcement_id) USING btree (user_id, announcement_id)
-
-foreign-key constraints
-    announcement_notifications_announcement_id_fkey: FOREIGN KEY (announcement_id) REFERENCES announcements(id) ON UPDATE CASCADE ON DELETE CASCADE
-    announcement_notifications_user_id_fkey: FOREIGN KEY (user_id) REFERENCES users(user_id) ON UPDATE CASCADE ON DELETE CASCADE
