--- database/tables/announcements.pg
@@ -1,15 +1,0 @@
-columns
-    date: timestamp with time zone not null default now()
-    directory: text not null
-    for_students: boolean not null default false
-    id: bigint not null default nextval('announcements_id_seq'::regclass)
-    order_by: integer not null default 0
-    title: text not null
-    uuid: uuid not null
-
-indexes
-    announcements_pkey: PRIMARY KEY (id) USING btree (id)
-    announcements_uuid_key: UNIQUE (uuid) USING btree (uuid)
-
-referenced by
-    announcement_notifications: FOREIGN KEY (announcement_id) REFERENCES announcements(id) ON UPDATE CASCADE ON DELETE CASCADE
