+++ database/tables/news_item_notifications.pg
@@ -1,0 +1,12 @@
+columns
+    news_item_id: bigint not null
+    id: bigint not null default nextval('news_item_notifications_id_seq'::regclass)
+    user_id: bigint not null
+
+indexes
+    news_item_notifications_pkey: PRIMARY KEY (id) USING btree (id)
+    news_item_notifications_user_id_news_item_id_key: UNIQUE (user_id, news_item_id) USING btree (user_id, news_item_id)
+
+foreign-key constraints
+    news_item_notifications_news_item_id_fkey: FOREIGN KEY (news_item_id) REFERENCES news_items(id) ON UPDATE CASCADE ON DELETE CASCADE
+    news_item_notifications_user_id_fkey: FOREIGN KEY (user_id) REFERENCES users(user_id) ON UPDATE CASCADE ON DELETE CASCADE
