+++ database/tables/news_items.pg
@@ -1,0 +1,15 @@
+columns
+    date: timestamp with time zone not null default now()
+    directory: text not null
+    for_students: boolean not null default false
+    id: bigint not null default nextval('news_items_id_seq'::regclass)
+    order_by: integer not null default 0
+    title: text not null
+    uuid: uuid not null
+
+indexes
+    news_items_pkey: PRIMARY KEY (id) USING btree (id)
+    news_items_uuid_key: UNIQUE (uuid) USING btree (uuid)
+
+referenced by
+    news_item_notifications: FOREIGN KEY (news_item_id) REFERENCES news_items(id) ON UPDATE CASCADE ON DELETE CASCADE
