+++ middlewares/authn.sql
@@ -3,7 +3,7 @@ SELECT
     to_jsonb(u.*) AS user,
     to_jsonb(i.*) AS institution,
     (adm.id IS NOT NULL) AS is_administrator,
-    (SELECT count(*) FROM announcement_notifications WHERE user_id = $user_id) AS announcement_notification_count
+    (SELECT count(*) FROM news_item_notifications WHERE user_id = $user_id) AS news_item_notification_count
 FROM
     users AS u
     LEFT JOIN administrators AS adm ON (adm.user_id = u.user_id)
