+++ news_items/002_in_browser_editing/info.json
