+++ pages/news_items/news_items.ejs
@@ -5,32 +5,32 @@
     <link href="/stylesheets/issues.css" rel="stylesheet" />
   </head>
   <body>
-    <%- include('../partials/navbar', {navPage: 'announcements'}); %>
+    <%- include('../partials/navbar', {navPage: 'news_items'}); %>
     <div id="content" class="container">
       <%- include('../partials/navtabs'); %>
 
       <div class="card mb-4">
         <div class="card-header bg-primary text-white d-flex">
-          Announcements
+          News
         </div>
 
-        <% if (rows.length == 0) { %>
+        <% if (news_items.length == 0) { %>
         <div class="card-body">
-          <div class="text-center text-muted">No announcements</div>
+          <div class="text-center text-muted">No news_items</div>
         </div>
         <% } else { %>
         <div class="table-responsive">
           <table class="table table-hover table-striped">
             <tbody>
-              <% rows.forEach(row => { %>
+              <% news_items.forEach(news_item => { %>
               <tr>
                 <td class="align-middle" style="width: 1%; white-space: nowrap;">
-                  <%= row.formatted_date %>
+                  <%= news_item.formatted_date %>
                 </td>
                 <td class="align-middle">
-                  <a href="<%= urlPrefix %>/announcement/<%= row.id %>/">
-                    <%= row.title %>
-                    <% if (row.unread) { %>
+                  <a href="<%= urlPrefix %>/news_item/<%= news_item.id %>/">
+                    <%= news_item.title %>
+                    <% if (news_item.unread) { %>
                     <span class="badge badge-danger ml-2">Unread</span>
                     <% } %>
                   </a>
