+++ pages/partials/navbar.ejs
@@ -51,8 +51,8 @@ if (typeof authz_data !== 'undefined' && authz_data.authn_has_instructor_view) {
     <li class="nav-item dropdown mb-2 mb-md-0 mr-2 <% if (navPage == "effective") { %>active<% } %>">
         <a class="nav-link dropdown-toggle" href="#" id="navbarDropdown" role="button" data-toggle="dropdown" aria-haspopup="true" aria-expanded="false">
             <%= displayedName %>
-            <% if (announcement_notification_count > 0) { %>
-              <span class="badge badge-pill badge-danger"><%= announcement_notification_count %></span>
+            <% if (news_item_notification_count > 0) { %>
+              <span class="badge badge-pill badge-danger"><%= news_item_notification_count %></span>
             <% } %>
         </a>
         <div class="dropdown-menu" aria-labelledby="navbarDropdown">
@@ -60,10 +60,10 @@ if (typeof authz_data !== 'undefined' && authz_data.authn_has_instructor_view) {
             <% if (typeof authz_data !== 'undefined' && authz_data.authn_has_instructor_view) { %>
             <a class="dropdown-item" href="<%= urlPrefix %>/effectiveUser">Change effective user</a>
             <% } %>
-            <a class="dropdown-item" href="<%= urlPrefix %>/announcements">
-              Announcements
-              <% if (announcement_notification_count > 0) { %>
-                <span class="badge badge-pill badge-danger"><%= announcement_notification_count %></span>
+            <a class="dropdown-item" href="<%= urlPrefix %>/news_items">
+              News
+              <% if (news_item_notification_count > 0) { %>
+                <span class="badge badge-pill badge-danger"><%= news_item_notification_count %></span>
               <% } %>
             </a>
         </div>
