+++ schemas/schemas/infoNewsItem.json
@@ -1,7 +1,7 @@
 {
     "$schema": "http://json-schema.org/draft-04/schema#",
-    "title": "Announcement Info",
-    "description": "Info files for announcements.",
+    "title": "News Item Info",
+    "description": "Info files for news items.",
     "type": "object",
     "additionalProperties": false,
     "required": [
@@ -16,11 +16,11 @@
             "pattern": "^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$"
         },
         "title": {
-            "description": "The title of the announcement.",
+            "description": "The title of the news item.",
             "type": "string"
         },
         "for_students": {
-            "description": "Whether the announcement should be shown to students.",
+            "description": "Whether the news item should be shown to students.",
             "type": "boolean"
         }
     }
