+++ server.js
@@ -29,7 +29,7 @@ const assessment = require('./lib/assessment');
 const sqldb = require('@prairielearn/prairielib/sql-db');
 const migrations = require('./migrations');
 const sprocs = require('./sprocs');
-const announcements = require('./announcements');
+const news_items = require('./news_items');
 const cron = require('./cron');
 const redis = require('./lib/redis');
 const socketServer = require('./lib/socket-server');
@@ -239,8 +239,8 @@ app.use('/pl/settings', require('./pages/userSettings/userSettings'));
 app.use('/pl/enroll', require('./pages/enroll/enroll'));
 app.use('/pl/logout', require('./pages/authLogout/authLogout'));
 app.use('/pl/password', require('./pages/authPassword/authPassword'));
-app.use('/pl/announcements', require('./pages/announcements/announcements.js'));
-app.use('/pl/announcement', require('./pages/announcement/announcement.js'));
+app.use('/pl/news_items', require('./pages/news_items/news_items.js'));
+app.use('/pl/news_item', require('./pages/news_item/news_item.js'));
 
 // dev-mode pages are mounted for both out-of-course access (here) and within-course access (see below)
 if (config.devMode) {
@@ -412,9 +412,9 @@ app.use('/pl/course_instance/:course_instance_id/instructor/loadFromDisk', requi
 app.use('/pl/course_instance/:course_instance_id/instructor/edit_error', require('./pages/editError/editError'));
 
 
-// course instance - announcements
-app.use('/pl/course_instance/:course_instance_id/instructor/announcements', require('./pages/announcements/announcements.js'));
-app.use('/pl/course_instance/:course_instance_id/instructor/announcement', require('./pages/announcement/announcement.js'));
+// course instance - news_items
+app.use('/pl/course_instance/:course_instance_id/instructor/news_items', require('./pages/news_items/news_items.js'));
+app.use('/pl/course_instance/:course_instance_id/instructor/news_item', require('./pages/news_item/news_item.js'));
 
 
 // course instance - course admin pages
@@ -590,9 +590,9 @@ if (config.devMode) {
     app.use('/pl/course_instance/:course_instance_id/jobSequence', require('./pages/instructorJobSequence/instructorJobSequence'));
 }
 
-// student - announcements
-app.use('/pl/course_instance/:course_instance_id/announcements', require('./pages/announcements/announcements.js'));
-app.use('/pl/course_instance/:course_instance_id/announcement', require('./pages/announcement/announcement.js'));
+// student - news_items
+app.use('/pl/course_instance/:course_instance_id/news_items', require('./pages/news_items/news_items.js'));
+app.use('/pl/course_instance/:course_instance_id/news_item', require('./pages/news_item/news_item.js'));
 
 
 // Allow access to effectiveUser as a Student page, but only for users have authn (not authz) as Instructor
@@ -690,9 +690,9 @@ app.use('/pl/course/:course_id/question/:question_id/file_view', [
 app.use('/pl/course/:course_id/question/:question_id/file_download', require('./pages/instructorFileDownload/instructorFileDownload'));
 
 
-// course - announcements
-app.use('/pl/course/:course_id/announcements', require('./pages/announcements/announcements.js'));
-app.use('/pl/course/:course_id/announcement', require('./pages/announcement/announcement.js'));
+// course - news_items
+app.use('/pl/course/:course_id/news_items', require('./pages/news_items/news_items.js'));
+app.use('/pl/course/:course_id/news_item', require('./pages/news_item/news_item.js'));
 
 
 app.use('/pl/course/:course_id/file_transfer', [
@@ -907,7 +907,7 @@ if (config.startServer) {
             }
         },
         function(callback) {
-            announcements.init(function(err) {
+            news_items.init(function(err) {
                 if (ERR(err, callback)) return;
                 callback(null);
             });
